refactor: reorganize workspace entity imports and remove deprecated files
- Updated import paths for various workspace entities to reflect new directory structure.
- Removed obsolete workspace entity files including mkt-attribute, mkt-combo, mkt-combo-variant, mkt-contract, mkt-customer, mkt-customer-tag, mkt-order, mkt-product, mkt-tag, mkt-template, mkt-value, mkt-variant, and mkt-variant-attribute.
- Enhanced code clarity by standardizing import formatting across multiple files.
