+++ packages/twenty-server/src/mkt-core/customer/objects/mkt-customer-tag.workspace-entity.ts
@@ -7,7 +7,6 @@ import { Relation } from 'src/engine/workspace-manager/workspace-sync-metadata/i
 
 import { SEARCH_VECTOR_FIELD } from 'src/engine/metadata-modules/constants/search-vector-field.constants';
 import { ActorMetadata } from 'src/engine/metadata-modules/field-metadata/composite-types/actor.composite-type';
-import { FieldMetadataComplexOption } from 'src/engine/metadata-modules/field-metadata/dtos/options.input';
 import { IndexType } from 'src/engine/metadata-modules/index-metadata/types/indexType.types';
 import { BaseWorkspaceEntity } from 'src/engine/twenty-orm/base.workspace-entity';
 import { WorkspaceDuplicateCriteria } from 'src/engine/twenty-orm/decorators/workspace-duplicate-criteria.decorator';
@@ -23,139 +22,53 @@ import {
   FieldTypeAndNameMetadata,
   getTsVectorColumnExpressionFromFields,
 } from 'src/engine/workspace-manager/workspace-sync-metadata/utils/get-ts-vector-column-expression.util';
-import { MktComboVariantWorkspaceEntity } from 'src/mkt-core/combo-variant/mkt-combo-variant.workspace-entity';
-import { MKT_COMBO_FIELD_IDS } from 'src/mkt-core/constants/mkt-field-ids';
+import { MKT_CUSTOMER_TAG_FIELD_IDS } from 'src/mkt-core/constants/mkt-field-ids';
 import { MKT_OBJECT_IDS } from 'src/mkt-core/constants/mkt-object-ids';
-import { MktOrderItemWorkspaceEntity } from 'src/mkt-core/order-item/mkt-order-item.workspace-entity';
+import { MktCustomerWorkspaceEntity } from 'src/mkt-core/customer/objects/mkt-customer.workspace-entity';
+import { MktTagWorkspaceEntity } from 'src/mkt-core/customer/objects/mkt-tag.workspace-entity';
 import { TimelineActivityWorkspaceEntity } from 'src/modules/timeline/standard-objects/timeline-activity.workspace-entity';
 import { WorkspaceMemberWorkspaceEntity } from 'src/modules/workspace-member/standard-objects/workspace-member.workspace-entity';
 
-const TABLE_COMBO_NAME = 'mktCombo';
+const TABLE_CUSTOMER_TAG_NAME = 'mktCustomerTag';
 const NAME_FIELD_NAME = 'name';
-const DESCRIPTION_FIELD_NAME = 'description';
 
-export const SEARCH_FIELDS_FOR_MKT_COMBO: FieldTypeAndNameMetadata[] = [
+export const SEARCH_FIELDS_FOR_MKT_CUSTOMER_TAG: FieldTypeAndNameMetadata[] = [
   { name: NAME_FIELD_NAME, type: FieldMetadataType.TEXT },
-  { name: DESCRIPTION_FIELD_NAME, type: FieldMetadataType.TEXT },
-];
-
-export enum MKT_COMBO_STATUS {
-  ACTIVE = 'ACTIVE',
-  INACTIVE = 'INACTIVE',
-  DELETED = 'DELETED',
-}
-
-export const MKT_COMBO_STATUS_OPTIONS: FieldMetadataComplexOption[] = [
-  {
-    value: MKT_COMBO_STATUS.ACTIVE,
-    label: 'Active',
-    position: 0,
-    color: 'blue',
-  },
-  {
-    value: MKT_COMBO_STATUS.INACTIVE,
-    label: 'Inactive',
-    position: 1,
-    color: 'purple',
-  },
-  {
-    value: MKT_COMBO_STATUS.DELETED,
-    label: 'Deleted',
-    position: 2,
-    color: 'green',
-  },
 ];
 
 @WorkspaceEntity({
-  standardId: MKT_OBJECT_IDS.mktCombo,
-  namePlural: `${TABLE_COMBO_NAME}s`,
-  labelSingular: msg`Combo`,
-  labelPlural: msg`Combos`,
-  description: msg`Combo entity for catalog`,
+  standardId: MKT_OBJECT_IDS.mktCustomerTag,
+  namePlural: `${TABLE_CUSTOMER_TAG_NAME}s`,
+  labelSingular: msg`Customer Tag`,
+  labelPlural: msg`Customer Tags`,
+  description: msg`Customer Tag entity for catalog`,
   icon: 'IconBox',
-  labelIdentifierStandardId: MKT_COMBO_FIELD_IDS.name,
+  labelIdentifierStandardId: MKT_CUSTOMER_TAG_FIELD_IDS.name,
 })
-@WorkspaceDuplicateCriteria([['name'], ['description']])
+@WorkspaceDuplicateCriteria([['mktCustomerId'], ['mktTagId']])
 @WorkspaceIsSearchable()
-export class MktComboWorkspaceEntity extends BaseWorkspaceEntity {
+export class MktCustomerTagWorkspaceEntity extends BaseWorkspaceEntity {
   @WorkspaceField({
-    standardId: MKT_COMBO_FIELD_IDS.name,
+    standardId: MKT_CUSTOMER_TAG_FIELD_IDS.name,
     type: FieldMetadataType.TEXT,
-    label: msg`Combo Name`,
-    description: msg`Combo name`,
+    label: msg`Customer Tag Name`,
+    description: msg`Customer Tag name`,
     icon: 'IconFileText',
   })
   name: string;
 
   @WorkspaceField({
-    standardId: MKT_COMBO_FIELD_IDS.status,
-    type: FieldMetadataType.SELECT,
-    label: msg`Combo Status`,
-    description: msg`Combo status (active, inactive, deleted)`,
-    icon: 'IconTags',
-    options: MKT_COMBO_STATUS_OPTIONS,
-  })
-  @WorkspaceIsNullable()
-  status: MKT_COMBO_STATUS;
-
-  @WorkspaceField({
-    standardId: MKT_COMBO_FIELD_IDS.description,
-    type: FieldMetadataType.TEXT,
-    label: msg`Combo Description`,
-    description: msg`Combo description`,
-    icon: 'IconBarcode',
-  })
-  @WorkspaceIsNullable()
-  description?: string;
-
-  @WorkspaceField({
-    standardId: MKT_COMBO_FIELD_IDS.price,
-    type: FieldMetadataType.NUMBER,
-    label: msg`Price`,
-    description: msg`Combo price`,
-    icon: 'IconClock',
-  })
-  @WorkspaceIsNullable()
-  price?: number;
-
-  @WorkspaceRelation({
-    standardId: MKT_COMBO_FIELD_IDS.mktComboVariants,
-    type: RelationType.ONE_TO_MANY,
-    label: msg`Combo Variants`,
-    description: msg`Combo variants`,
-    icon: 'IconBox',
-    inverseSideTarget: () => MktComboVariantWorkspaceEntity,
-    inverseSideFieldKey: 'mktCombo',
-    onDelete: RelationOnDeleteAction.SET_NULL,
-  })
-  @WorkspaceIsNullable()
-  mktComboVariants?: Relation<MktComboVariantWorkspaceEntity[]>;
-
-  @WorkspaceRelation({
-    standardId: MKT_COMBO_FIELD_IDS.mktOrderItems,
-    type: RelationType.ONE_TO_MANY,
-    label: msg`Order Items`,
-    description: msg`Combo order items`,
-    icon: 'IconDeviceDesktop',
-    inverseSideTarget: () => MktOrderItemWorkspaceEntity,
-    inverseSideFieldKey: 'mktCombo',
-    onDelete: RelationOnDeleteAction.CASCADE,
-  })
-  @WorkspaceIsNullable()
-  mktOrderItems?: Relation<MktOrderItemWorkspaceEntity[]>;
-
-  @WorkspaceField({
-    standardId: MKT_COMBO_FIELD_IDS.position,
+    standardId: MKT_CUSTOMER_TAG_FIELD_IDS.position,
     type: FieldMetadataType.POSITION,
     label: msg`Position`,
-    description: msg`Combo position`,
-    icon: 'IconFileText',
+    description: msg`Position in list`,
+    icon: 'IconHierarchy',
   })
   @WorkspaceIsNullable()
   position: number;
 
   @WorkspaceField({
-    standardId: MKT_COMBO_FIELD_IDS.createdBy,
+    standardId: MKT_CUSTOMER_TAG_FIELD_IDS.createdBy,
     type: FieldMetadataType.ACTOR,
     label: msg`Created by`,
     icon: 'IconCreativeCommonsSa',
@@ -164,13 +77,41 @@ export class MktComboWorkspaceEntity extends BaseWorkspaceEntity {
   createdBy: ActorMetadata;
 
   @WorkspaceRelation({
-    standardId: MKT_COMBO_FIELD_IDS.accountOwner,
+    standardId: MKT_CUSTOMER_TAG_FIELD_IDS.mktCustomer,
+    type: RelationType.MANY_TO_ONE,
+    label: msg`Customer`,
+    description: msg`Customer tag customer`,
+    icon: 'IconUser',
+    inverseSideTarget: () => MktCustomerWorkspaceEntity,
+    inverseSideFieldKey: 'mktCustomerTags',
+    onDelete: RelationOnDeleteAction.CASCADE,
+  })
+  mktCustomer: Relation<MktCustomerWorkspaceEntity> | null;
+  @WorkspaceJoinColumn('mktCustomer')
+  mktCustomerId: string | null;
+
+  @WorkspaceRelation({
+    standardId: MKT_CUSTOMER_TAG_FIELD_IDS.mktTag,
+    type: RelationType.MANY_TO_ONE,
+    label: msg`Tag`,
+    description: msg`Customer tag tag`,
+    icon: 'IconTag',
+    inverseSideTarget: () => MktTagWorkspaceEntity,
+    inverseSideFieldKey: 'mktCustomerTags',
+    onDelete: RelationOnDeleteAction.CASCADE,
+  })
+  mktTag: Relation<MktTagWorkspaceEntity> | null;
+  @WorkspaceJoinColumn('mktTag')
+  mktTagId: string | null;
+
+  @WorkspaceRelation({
+    standardId: MKT_CUSTOMER_TAG_FIELD_IDS.accountOwner,
     type: RelationType.MANY_TO_ONE,
     label: msg`Account Owner`,
-    description: msg`Your team member responsible for managing the combo`,
+    description: msg`Your team member responsible for managing the customer tag`,
     icon: 'IconUserCircle',
     inverseSideTarget: () => WorkspaceMemberWorkspaceEntity,
-    inverseSideFieldKey: 'accountOwnerForMktCombos',
+    inverseSideFieldKey: 'accountOwnerForMktCustomerTags',
     onDelete: RelationOnDeleteAction.SET_NULL,
   })
   @WorkspaceIsNullable()
@@ -180,13 +121,13 @@ export class MktComboWorkspaceEntity extends BaseWorkspaceEntity {
   accountOwnerId: string | null;
 
   @WorkspaceRelation({
-    standardId: MKT_COMBO_FIELD_IDS.timelineActivities,
+    standardId: MKT_CUSTOMER_TAG_FIELD_IDS.timelineActivities,
     type: RelationType.ONE_TO_MANY,
     label: msg`Timeline Activities`,
-    description: msg`Timeline Activities linked to the combo`,
+    description: msg`Timeline Activities linked to the customer tag`,
     icon: 'IconIconTimelineEvent',
     inverseSideTarget: () => TimelineActivityWorkspaceEntity,
-    inverseSideFieldKey: 'mktCombo',
+    inverseSideFieldKey: 'mktCustomerTag',
     onDelete: RelationOnDeleteAction.CASCADE,
   })
   @WorkspaceIsNullable()
@@ -194,14 +135,14 @@ export class MktComboWorkspaceEntity extends BaseWorkspaceEntity {
   timelineActivities: Relation<TimelineActivityWorkspaceEntity[]>;
 
   @WorkspaceField({
-    standardId: MKT_COMBO_FIELD_IDS.searchVector,
+    standardId: MKT_CUSTOMER_TAG_FIELD_IDS.searchVector,
     type: FieldMetadataType.TS_VECTOR,
     label: SEARCH_VECTOR_FIELD.label,
     description: SEARCH_VECTOR_FIELD.description,
     icon: 'IconUser',
     generatedType: 'STORED',
     asExpression: getTsVectorColumnExpressionFromFields(
-      SEARCH_FIELDS_FOR_MKT_COMBO,
+      SEARCH_FIELDS_FOR_MKT_CUSTOMER_TAG,
     ),
   })
   @WorkspaceIsNullable()
