+++ packages/twenty-server/src/mkt-core/customer/objects/mkt-customer.workspace-entity.ts
@@ -23,7 +23,6 @@ import {
 } from 'src/engine/workspace-manager/workspace-sync-metadata/utils/get-ts-vector-column-expression.util';
 import { MKT_CUSTOMER_FIELD_IDS } from 'src/mkt-core/constants/mkt-field-ids';
 import { MKT_OBJECT_IDS } from 'src/mkt-core/constants/mkt-object-ids';
-import { MktCustomerTagWorkspaceEntity } from 'src/mkt-core/customer-tag/mkt-customer-tag.workspace-entity';
 import {
   MKT_CUSTOMER_COMPANY_SIZE,
   MKT_CUSTOMER_COMPANY_SIZE_OPTIONS,
@@ -40,6 +39,7 @@ import {
   MKT_CUSTOMER_TYPE,
   MKT_CUSTOMER_TYPE_OPTIONS,
 } from 'src/mkt-core/customer/constants/mkt-customer.constant';
+import { MktCustomerTagWorkspaceEntity } from 'src/mkt-core/customer/objects/mkt-customer-tag.workspace-entity';
 import { MktLicenseWorkspaceEntity } from 'src/mkt-core/license/mkt-license.workspace-entity';
 import { TimelineActivityWorkspaceEntity } from 'src/modules/timeline/standard-objects/timeline-activity.workspace-entity';
 import { WorkspaceMemberWorkspaceEntity } from 'src/modules/workspace-member/standard-objects/workspace-member.workspace-entity';
