+++ packages/twenty-server/src/mkt-core/invoice/s-invoice.integration.service.ts
@@ -7,8 +7,8 @@ import axios, { AxiosInstance } from 'axios';
 import { ScopedWorkspaceContextFactory } from 'src/engine/twenty-orm/factories/scoped-workspace-context.factory';
 import { TwentyORMGlobalManager } from 'src/engine/twenty-orm/twenty-orm-global.manager';
 import { MKT_INVOICE_STATUS } from 'src/mkt-core/invoice/objects/mkt-invoice.workspace-entity';
-import { MktOrderItemWorkspaceEntity } from 'src/mkt-core/order-item/mkt-order-item.workspace-entity';
-import { MktOrderWorkspaceEntity } from 'src/mkt-core/order/mkt-order.workspace-entity';
+import { MktOrderItemWorkspaceEntity } from 'src/mkt-core/order/objects/mkt-order-item.workspace-entity';
+import { MktOrderWorkspaceEntity } from 'src/mkt-core/order/objects/mkt-order.workspace-entity';
 
 type CreateInvoiceResponse = {
   transactionUuid?: string;
