+++ packages/twenty-server/src/mkt-core/product/objects/mkt-combo.workspace-entity.ts
@@ -23,117 +23,139 @@ import {
   FieldTypeAndNameMetadata,
   getTsVectorColumnExpressionFromFields,
 } from 'src/engine/workspace-manager/workspace-sync-metadata/utils/get-ts-vector-column-expression.util';
-import { MKT_CONTRACT_FIELD_IDS } from 'src/mkt-core/constants/mkt-field-ids';
+import { MKT_COMBO_FIELD_IDS } from 'src/mkt-core/constants/mkt-field-ids';
 import { MKT_OBJECT_IDS } from 'src/mkt-core/constants/mkt-object-ids';
-import { MktOrderWorkspaceEntity } from 'src/mkt-core/order/mkt-order.workspace-entity';
-// import { TimelineActivityWorkspaceEntity } from 'src/modules/timeline/standard-objects/timeline-activity.workspace-entity';
+import { MktOrderItemWorkspaceEntity } from 'src/mkt-core/order/objects/mkt-order-item.workspace-entity';
+import { MktComboVariantWorkspaceEntity } from 'src/mkt-core/product/objects/mkt-combo-variant.workspace-entity';
+import { TimelineActivityWorkspaceEntity } from 'src/modules/timeline/standard-objects/timeline-activity.workspace-entity';
 import { WorkspaceMemberWorkspaceEntity } from 'src/modules/workspace-member/standard-objects/workspace-member.workspace-entity';
 
-const TABLE_CONTRACT_NAME = 'mktContract';
+const TABLE_COMBO_NAME = 'mktCombo';
 const NAME_FIELD_NAME = 'name';
+const DESCRIPTION_FIELD_NAME = 'description';
 
-export const SEARCH_FIELDS_FOR_MKT_CONTRACT: FieldTypeAndNameMetadata[] = [
+export const SEARCH_FIELDS_FOR_MKT_COMBO: FieldTypeAndNameMetadata[] = [
   { name: NAME_FIELD_NAME, type: FieldMetadataType.TEXT },
+  { name: DESCRIPTION_FIELD_NAME, type: FieldMetadataType.TEXT },
 ];
 
-export enum MKT_CONTRACT_STATUS {
-  ACTIVE = 'active',
-  INACTIVE = 'inactive',
-  EXPIRED = 'expired',
-  REVOKED = 'revoked',
+export enum MKT_COMBO_STATUS {
+  ACTIVE = 'ACTIVE',
+  INACTIVE = 'INACTIVE',
+  DELETED = 'DELETED',
 }
 
-export const MKT_CONTRACT_STATUS_OPTIONS: FieldMetadataComplexOption[] = [
+export const MKT_COMBO_STATUS_OPTIONS: FieldMetadataComplexOption[] = [
   {
-    value: MKT_CONTRACT_STATUS.ACTIVE,
+    value: MKT_COMBO_STATUS.ACTIVE,
     label: 'Active',
     position: 0,
     color: 'blue',
   },
   {
-    value: MKT_CONTRACT_STATUS.INACTIVE,
+    value: MKT_COMBO_STATUS.INACTIVE,
     label: 'Inactive',
     position: 1,
     color: 'purple',
   },
   {
-    value: MKT_CONTRACT_STATUS.EXPIRED,
-    label: 'Expired',
+    value: MKT_COMBO_STATUS.DELETED,
+    label: 'Deleted',
     position: 2,
     color: 'green',
-  },
-  {
-    value: MKT_CONTRACT_STATUS.REVOKED,
-    label: 'Revoked',
-    position: 3,
-    color: 'orange',
   },
 ];
 
 @WorkspaceEntity({
-  standardId: MKT_OBJECT_IDS.mktContract,
-  namePlural: `${TABLE_CONTRACT_NAME}s`,
-  labelSingular: msg`Contract`,
-  labelPlural: msg`Contracts`,
-  description: msg`Contract entity for catalog`,
+  standardId: MKT_OBJECT_IDS.mktCombo,
+  namePlural: `${TABLE_COMBO_NAME}s`,
+  labelSingular: msg`Combo`,
+  labelPlural: msg`Combos`,
+  description: msg`Combo entity for catalog`,
   icon: 'IconBox',
-  labelIdentifierStandardId: MKT_CONTRACT_FIELD_IDS.name,
+  labelIdentifierStandardId: MKT_COMBO_FIELD_IDS.name,
 })
-@WorkspaceDuplicateCriteria([['name']])
+@WorkspaceDuplicateCriteria([['name'], ['description']])
 @WorkspaceIsSearchable()
-export class MktContractWorkspaceEntity extends BaseWorkspaceEntity {
+export class MktComboWorkspaceEntity extends BaseWorkspaceEntity {
   @WorkspaceField({
-    standardId: MKT_CONTRACT_FIELD_IDS.name,
+    standardId: MKT_COMBO_FIELD_IDS.name,
     type: FieldMetadataType.TEXT,
-    label: msg`License Name`,
-    description: msg`License name`,
+    label: msg`Combo Name`,
+    description: msg`Combo name`,
     icon: 'IconFileText',
   })
   name: string;
 
   @WorkspaceField({
-    standardId: MKT_CONTRACT_FIELD_IDS.status,
+    standardId: MKT_COMBO_FIELD_IDS.status,
     type: FieldMetadataType.SELECT,
-    label: msg`License Status`,
-    description: msg`License status (active, inactive, expired, revoked)`,
+    label: msg`Combo Status`,
+    description: msg`Combo status (active, inactive, deleted)`,
     icon: 'IconTags',
-    options: MKT_CONTRACT_STATUS_OPTIONS,
+    options: MKT_COMBO_STATUS_OPTIONS,
   })
   @WorkspaceIsNullable()
-  status: MKT_CONTRACT_STATUS;
+  status: MKT_COMBO_STATUS;
 
   @WorkspaceField({
-    standardId: MKT_CONTRACT_FIELD_IDS.startDate,
+    standardId: MKT_COMBO_FIELD_IDS.description,
     type: FieldMetadataType.TEXT,
-    label: msg`Start Date`,
-    description: msg`Contract start date`,
+    label: msg`Combo Description`,
+    description: msg`Combo description`,
     icon: 'IconBarcode',
   })
   @WorkspaceIsNullable()
-  startDate?: string;
+  description?: string;
 
   @WorkspaceField({
-    standardId: MKT_CONTRACT_FIELD_IDS.endDate,
-    type: FieldMetadataType.TEXT,
-    label: msg`End Date`,
-    description: msg`Contract end date`,
+    standardId: MKT_COMBO_FIELD_IDS.price,
+    type: FieldMetadataType.NUMBER,
+    label: msg`Price`,
+    description: msg`Combo price`,
     icon: 'IconClock',
   })
   @WorkspaceIsNullable()
-  endDate?: string;
+  price?: number;
+
+  @WorkspaceRelation({
+    standardId: MKT_COMBO_FIELD_IDS.mktComboVariants,
+    type: RelationType.ONE_TO_MANY,
+    label: msg`Combo Variants`,
+    description: msg`Combo variants`,
+    icon: 'IconBox',
+    inverseSideTarget: () => MktComboVariantWorkspaceEntity,
+    inverseSideFieldKey: 'mktCombo',
+    onDelete: RelationOnDeleteAction.SET_NULL,
+  })
+  @WorkspaceIsNullable()
+  mktComboVariants?: Relation<MktComboVariantWorkspaceEntity[]>;
+
+  @WorkspaceRelation({
+    standardId: MKT_COMBO_FIELD_IDS.mktOrderItems,
+    type: RelationType.ONE_TO_MANY,
+    label: msg`Order Items`,
+    description: msg`Combo order items`,
+    icon: 'IconDeviceDesktop',
+    inverseSideTarget: () => MktOrderItemWorkspaceEntity,
+    inverseSideFieldKey: 'mktCombo',
+    onDelete: RelationOnDeleteAction.CASCADE,
+  })
+  @WorkspaceIsNullable()
+  mktOrderItems?: Relation<MktOrderItemWorkspaceEntity[]>;
 
   @WorkspaceField({
-    standardId: MKT_CONTRACT_FIELD_IDS.position,
+    standardId: MKT_COMBO_FIELD_IDS.position,
     type: FieldMetadataType.POSITION,
     label: msg`Position`,
-    description: msg`Position in list`,
-    icon: 'IconHierarchy',
+    description: msg`Combo position`,
+    icon: 'IconFileText',
   })
   @WorkspaceIsNullable()
   position: number;
 
   @WorkspaceField({
-    standardId: MKT_CONTRACT_FIELD_IDS.createdBy,
+    standardId: MKT_COMBO_FIELD_IDS.createdBy,
     type: FieldMetadataType.ACTOR,
     label: msg`Created by`,
     icon: 'IconCreativeCommonsSa',
@@ -142,30 +164,13 @@ export class MktContractWorkspaceEntity extends BaseWorkspaceEntity {
   createdBy: ActorMetadata;
 
   @WorkspaceRelation({
-    standardId: MKT_CONTRACT_FIELD_IDS.mktOrder,
-    type: RelationType.MANY_TO_ONE,
-    label: msg`Order`,
-    description: msg`Order that this contract belongs to`,
-    icon: 'IconShoppingCart',
-    inverseSideTarget: () => MktOrderWorkspaceEntity,
-    inverseSideFieldKey: 'mktContracts',
-    onDelete: RelationOnDeleteAction.CASCADE,
-  })
-  @WorkspaceIsNullable()
-  mktOrder: Relation<MktOrderWorkspaceEntity> | null;
-
-  @WorkspaceJoinColumn('mktOrder')
-  @WorkspaceIsNullable()
-  mktOrderId: string | null;
-
-  @WorkspaceRelation({
-    standardId: MKT_CONTRACT_FIELD_IDS.accountOwner,
+    standardId: MKT_COMBO_FIELD_IDS.accountOwner,
     type: RelationType.MANY_TO_ONE,
     label: msg`Account Owner`,
-    description: msg`Your team member responsible for managing the contract`,
+    description: msg`Your team member responsible for managing the combo`,
     icon: 'IconUserCircle',
     inverseSideTarget: () => WorkspaceMemberWorkspaceEntity,
-    inverseSideFieldKey: 'accountOwnerForMktContracts',
+    inverseSideFieldKey: 'accountOwnerForMktCombos',
     onDelete: RelationOnDeleteAction.SET_NULL,
   })
   @WorkspaceIsNullable()
@@ -174,30 +179,29 @@ export class MktContractWorkspaceEntity extends BaseWorkspaceEntity {
   @WorkspaceJoinColumn('accountOwner')
   accountOwnerId: string | null;
 
-  // Temporarily commented out due to relation conflicts with TimelineActivity
-  // @WorkspaceRelation({
-  //   standardId: MKT_CONTRACT_FIELD_IDS.timelineActivities,
-  //   type: RelationType.ONE_TO_MANY,
-  //   label: msg`Timeline Activities`,
-  //   description: msg`Timeline Activities linked to the contract`,
-  //   icon: 'IconIconTimelineEvent',
-  //   inverseSideTarget: () => TimelineActivityWorkspaceEntity,
-  //   inverseSideFieldKey: 'mktContract',
-  //   onDelete: RelationOnDeleteAction.CASCADE,
-  // })
-  // @WorkspaceIsNullable()
-  // @WorkspaceIsSystem()
-  // timelineActivities: Relation<TimelineActivityWorkspaceEntity[]>;
+  @WorkspaceRelation({
+    standardId: MKT_COMBO_FIELD_IDS.timelineActivities,
+    type: RelationType.ONE_TO_MANY,
+    label: msg`Timeline Activities`,
+    description: msg`Timeline Activities linked to the combo`,
+    icon: 'IconIconTimelineEvent',
+    inverseSideTarget: () => TimelineActivityWorkspaceEntity,
+    inverseSideFieldKey: 'mktCombo',
+    onDelete: RelationOnDeleteAction.CASCADE,
+  })
+  @WorkspaceIsNullable()
+  @WorkspaceIsSystem()
+  timelineActivities: Relation<TimelineActivityWorkspaceEntity[]>;
 
   @WorkspaceField({
-    standardId: MKT_CONTRACT_FIELD_IDS.searchVector,
+    standardId: MKT_COMBO_FIELD_IDS.searchVector,
     type: FieldMetadataType.TS_VECTOR,
     label: SEARCH_VECTOR_FIELD.label,
     description: SEARCH_VECTOR_FIELD.description,
     icon: 'IconUser',
     generatedType: 'STORED',
     asExpression: getTsVectorColumnExpressionFromFields(
-      SEARCH_FIELDS_FOR_MKT_CONTRACT,
+      SEARCH_FIELDS_FOR_MKT_COMBO,
     ),
   })
   @WorkspaceIsNullable()
