+++ packages/twenty-server/src/mkt-core/product/objects/mkt-variant-attribute.workspace-entity.ts
@@ -1,13 +1,13 @@
 import { msg } from '@lingui/core/macro';
 import { FieldMetadataType } from 'twenty-shared/types';
-import { Relation } from 'typeorm';
 
+import { RelationOnDeleteAction } from 'src/engine/metadata-modules/field-metadata/interfaces/relation-on-delete-action.interface';
 import { RelationType } from 'src/engine/metadata-modules/field-metadata/interfaces/relation-type.interface';
+import { Relation } from 'src/engine/workspace-manager/workspace-sync-metadata/interfaces/relation.interface';
 
 import { SEARCH_VECTOR_FIELD } from 'src/engine/metadata-modules/constants/search-vector-field.constants';
 import { ActorMetadata } from 'src/engine/metadata-modules/field-metadata/composite-types/actor.composite-type';
 import { IndexType } from 'src/engine/metadata-modules/index-metadata/types/indexType.types';
-import { RelationOnDeleteAction } from 'src/engine/metadata-modules/relation-metadata/relation-on-delete-action.type';
 import { BaseWorkspaceEntity } from 'src/engine/twenty-orm/base.workspace-entity';
 import { WorkspaceEntity } from 'src/engine/twenty-orm/decorators/workspace-entity.decorator';
 import { WorkspaceFieldIndex } from 'src/engine/twenty-orm/decorators/workspace-field-index.decorator';
@@ -21,41 +21,41 @@ import {
   FieldTypeAndNameMetadata,
   getTsVectorColumnExpressionFromFields,
 } from 'src/engine/workspace-manager/workspace-sync-metadata/utils/get-ts-vector-column-expression.util';
-import { MktAttributeWorkspaceEntity } from 'src/mkt-core/attribute/mkt-attribute.workspace-entity';
-import { MKT_VALUE_FIELD_IDS } from 'src/mkt-core/constants/mkt-field-ids';
+import { MKT_VARIANT_ATTRIBUTE_FIELD_IDS } from 'src/mkt-core/constants/mkt-field-ids';
 import { MKT_OBJECT_IDS } from 'src/mkt-core/constants/mkt-object-ids';
+import { MktAttributeWorkspaceEntity } from 'src/mkt-core/product/objects/mkt-attribute.workspace-entity';
+import { MktVariantWorkspaceEntity } from 'src/mkt-core/product/objects/mkt-variant.workspace-entity';
 import { TimelineActivityWorkspaceEntity } from 'src/modules/timeline/standard-objects/timeline-activity.workspace-entity';
 import { WorkspaceMemberWorkspaceEntity } from 'src/modules/workspace-member/standard-objects/workspace-member.workspace-entity';
 
-const VALUE_TABLE_NAME = 'mktValue';
-const VALUE_FIELD_NAME = 'name';
-
-export const SEARCH_FIELDS_FOR_MKT_VALUE: FieldTypeAndNameMetadata[] = [
-  { name: VALUE_FIELD_NAME, type: FieldMetadataType.TEXT },
+const TABLE_VARIANT_ATTRIBUTE_NAME = 'mktVariantAttribute';
+const NAME_FIELD_NAME = 'name';
+const SEARCH_FIELDS_FOR_MKT_VARIANT_ATTRIBUTE: FieldTypeAndNameMetadata[] = [
+  { name: NAME_FIELD_NAME, type: FieldMetadataType.TEXT },
 ];
 
 @WorkspaceEntity({
-  standardId: MKT_OBJECT_IDS.mktValue,
-  namePlural: `${VALUE_TABLE_NAME}s`,
-  labelSingular: msg`Value`,
-  labelPlural: msg`Values (Product Attribute)`,
-  description: msg`Attribute values (e.g., Color, Size, ...)`,
+  standardId: MKT_OBJECT_IDS.mktVariantAttribute,
+  namePlural: `${TABLE_VARIANT_ATTRIBUTE_NAME}s`,
+  labelSingular: msg`Variant Attribute`,
+  labelPlural: msg`Variant Attributes (Product)`,
+  description: msg`Assign attributes to variants`,
   icon: 'IconListDetails',
-  labelIdentifierStandardId: MKT_VALUE_FIELD_IDS.name,
+  labelIdentifierStandardId: MKT_VARIANT_ATTRIBUTE_FIELD_IDS.name,
 })
 @WorkspaceIsSearchable()
-export class MktValueWorkspaceEntity extends BaseWorkspaceEntity {
+export class MktVariantAttributeWorkspaceEntity extends BaseWorkspaceEntity {
   @WorkspaceField({
-    standardId: MKT_VALUE_FIELD_IDS.name,
+    standardId: MKT_VARIANT_ATTRIBUTE_FIELD_IDS.name,
     type: FieldMetadataType.TEXT,
     label: msg`Name`,
-    description: msg`Value name`,
+    description: msg`Name of the variant attribute`,
     icon: 'IconAbc',
   })
   name: string;
 
   @WorkspaceField({
-    standardId: MKT_VALUE_FIELD_IDS.position,
+    standardId: MKT_VARIANT_ATTRIBUTE_FIELD_IDS.position,
     type: FieldMetadataType.POSITION,
     label: msg`Position`,
     description: msg`Position in the list`,
@@ -65,22 +65,38 @@ export class MktValueWorkspaceEntity extends BaseWorkspaceEntity {
   position?: number;
 
   @WorkspaceField({
-    standardId: MKT_VALUE_FIELD_IDS.createdBy,
+    standardId: MKT_VARIANT_ATTRIBUTE_FIELD_IDS.createdBy,
     type: FieldMetadataType.ACTOR,
-    label: msg`Created By `,
-    description: msg`The creator of the value`,
+    label: msg`Created By`,
+    description: msg`The creator of the variant attribute`,
     icon: 'IconUserCircle',
   })
   createdBy: ActorMetadata;
 
   @WorkspaceRelation({
-    standardId: MKT_VALUE_FIELD_IDS.mktAttribute,
+    standardId: MKT_VARIANT_ATTRIBUTE_FIELD_IDS.mktVariant,
+    type: RelationType.MANY_TO_ONE,
+    label: msg`Variant`,
+    description: msg`Variant of the attribute`,
+    icon: 'IconList',
+    inverseSideTarget: () => MktVariantWorkspaceEntity,
+    inverseSideFieldKey: 'mktVariantAttributes',
+    onDelete: RelationOnDeleteAction.SET_NULL,
+  })
+  @WorkspaceIsNullable()
+  mktVariant: Relation<MktVariantWorkspaceEntity> | null;
+
+  @WorkspaceJoinColumn('mktVariant')
+  mktVariantId: string | null;
+
+  @WorkspaceRelation({
+    standardId: MKT_VARIANT_ATTRIBUTE_FIELD_IDS.mktAttribute,
     type: RelationType.MANY_TO_ONE,
     label: msg`Attribute`,
-    description: msg`Attribute for this value`,
-    icon: 'IconList',
+    description: msg`Attribute of the variant`,
+    icon: 'IconTag',
     inverseSideTarget: () => MktAttributeWorkspaceEntity,
-    inverseSideFieldKey: 'mktValues',
+    inverseSideFieldKey: 'mktVariantAttributes',
     onDelete: RelationOnDeleteAction.SET_NULL,
   })
   @WorkspaceIsNullable()
@@ -90,13 +106,13 @@ export class MktValueWorkspaceEntity extends BaseWorkspaceEntity {
   mktAttributeId: string | null;
 
   @WorkspaceRelation({
-    standardId: MKT_VALUE_FIELD_IDS.accountOwner,
+    standardId: MKT_VARIANT_ATTRIBUTE_FIELD_IDS.accountOwner,
     type: RelationType.MANY_TO_ONE,
     label: msg`Account Owner`,
-    description: msg`Your team member responsible for managing the value`,
+    description: msg`Your team member responsible for managing the variant attribute`,
     icon: 'IconUserCircle',
     inverseSideTarget: () => WorkspaceMemberWorkspaceEntity,
-    inverseSideFieldKey: 'accountOwnerForMktValues',
+    inverseSideFieldKey: 'accountOwnerForMktVariantAttributes',
     onDelete: RelationOnDeleteAction.SET_NULL,
   })
   @WorkspaceIsNullable()
@@ -106,13 +122,13 @@ export class MktValueWorkspaceEntity extends BaseWorkspaceEntity {
   accountOwnerId: string | null;
 
   @WorkspaceRelation({
-    standardId: MKT_VALUE_FIELD_IDS.timelineActivities,
+    standardId: MKT_VARIANT_ATTRIBUTE_FIELD_IDS.timelineActivities,
     type: RelationType.ONE_TO_MANY,
     label: msg`Timeline Activities`,
-    description: msg`Timeline Activities linked to the value`,
+    description: msg`Timeline Activities linked to the variant attribute`,
     icon: 'IconIconTimelineEvent',
     inverseSideTarget: () => TimelineActivityWorkspaceEntity,
-    inverseSideFieldKey: 'mktValue',
+    inverseSideFieldKey: 'mktVariantAttribute',
     onDelete: RelationOnDeleteAction.CASCADE,
   })
   @WorkspaceIsNullable()
@@ -120,14 +136,14 @@ export class MktValueWorkspaceEntity extends BaseWorkspaceEntity {
   timelineActivities: Relation<TimelineActivityWorkspaceEntity[]>;
 
   @WorkspaceField({
-    standardId: MKT_VALUE_FIELD_IDS.searchVector,
+    standardId: MKT_VARIANT_ATTRIBUTE_FIELD_IDS.searchVector,
     type: FieldMetadataType.TS_VECTOR,
     label: SEARCH_VECTOR_FIELD.label,
     description: SEARCH_VECTOR_FIELD.description,
     icon: 'IconUser',
     generatedType: 'STORED',
     asExpression: getTsVectorColumnExpressionFromFields(
-      SEARCH_FIELDS_FOR_MKT_VALUE,
+      SEARCH_FIELDS_FOR_MKT_VARIANT_ATTRIBUTE,
     ),
   })
   @WorkspaceIsNullable()
